+++ packages/core/plugin/src/withSentry.ts
@@ -3,6 +3,8 @@ import { createRunOncePlugin } from 'expo/config-plugins';
 
 import { bold, sdkPackage, warnOnce } from './utils';
 import { withSentryAndroid } from './withSentryAndroid';
+import type { SentryAndroidGradlePluginOptions } from './withSentryAndroidGradlePlugin';
+import { withSentryAndroidGradlePlugin } from './withSentryAndroidGradlePlugin';
 import { withSentryIOS } from './withSentryIOS';
 
 interface PluginProps {
@@ -10,6 +12,7 @@ interface PluginProps {
   project?: string;
   authToken?: string;
   url?: string;
+  experimental_android?: SentryAndroidGradlePluginOptions;
 }
 
 const withSentryPlugin: ConfigPlugin<PluginProps | void> = (config, props) => {
@@ -26,6 +29,14 @@ const withSentryPlugin: ConfigPlugin<PluginProps | void> = (config, props) => {
       cfg = withSentryAndroid(cfg, sentryProperties);
     } catch (e) {
       warnOnce(`There was a problem with configuring your native Android project: ${e}`);
+    }
+    // if `enableAndroidGradlePlugin` is provided configure the Sentry Android Gradle Plugin
+    if (props?.experimental_android && props?.experimental_android?.enableAndroidGradlePlugin) {
+      try {
+        cfg = withSentryAndroidGradlePlugin(cfg, props.experimental_android);
+      } catch (e) {
+        warnOnce(`There was a problem with configuring Sentry Android Gradle Plugin: ${e}`);
+      }
     }
     try {
       cfg = withSentryIOS(cfg, sentryProperties);